Split `selectors/index.ts` as it is becoming too big
